refactor(sync): split sync.ts orchestrator into cohesive modules under 300 SLOC
Eliminate the last (and largest) file-length exemption. sync.ts (1306 → 284)
keeps only runSync + a transitional re-export facade so existing importers
(webhook handlers, admin, bootstrap, sync.test.ts) are unchanged. Extracted
cohesive sibling modules, each well under the 300-SLOC cap:

  constants(3) label-vocab(95) upsert(120) control(90) repo-issues(104)
  repo-branches(69) repo-prs(116) bundle(168) edges(138) audit(88) tenants(96)

MAX_PAGES/WINDOW/NUM_SLOTS moved to constants.ts to avoid a circular import
through the facade. Pure moves, no logic change; 729 tests pass unchanged.

All 5 production file-length exemptions are now eliminated (0 remaining).
The 11 test-file exemptions stay (test exclusion is an upstream dev-core gate
change — is_exempt does exact-match, no glob; canonical script not editable here).
