+++ CLAUDE.md
@@ -42,7 +42,18 @@ Let:
 | `worker/src/graph/status.ts` | Graph status compute + server-side filter |
 | `worker/src/api/admin.ts` | `POST /admin/sync` (ADMIN_TOKEN-gated) |
 | `worker/src/api/version.ts` | `GET /api/version` |
-| `worker/src/sync/sync.ts` | Daily reconcile orchestrator + R2 audit write |
+| `worker/src/sync/sync.ts` | Daily reconcile orchestrator (`runSync`) + public facade re-exports |
+| `worker/src/sync/constants.ts` | Pagination/windowing constants (MAX_PAGES/WINDOW/NUM_SLOTS) |
+| `worker/src/sync/label-vocab.ts` | Label vocab + key helpers (extractFromLabels/canonicalKey/collectEdges) |
+| `worker/src/sync/upsert.ts` | Issue upsert SQL + prepareIssueUpsert (zk title redaction) |
+| `worker/src/sync/control.ts` | batchChunked + sync_control helpers (lock/halt/auth-failures) |
+| `worker/src/sync/repo-issues.ts` | Per-repo issue sync (standalone ISSUES_QUERY) |
+| `worker/src/sync/repo-branches.ts` | Branch sync (has_active_branch) |
+| `worker/src/sync/repo-prs.ts` | PR-state sync (applyPrState/UPSERT_PR_STATE_SQL) |
+| `worker/src/sync/bundle.ts` | Bundled per-repo fetch (issues+refs+PRs) — primary path |
+| `worker/src/sync/edges.ts` | Edge flush (pass 2) + closed-hop stub backfill |
+| `worker/src/sync/audit.ts` | Per-run audit summary (R2 + Workers Logs) |
+| `worker/src/sync/tenants.ts` | Phase-1 per-tenant repo discovery |
 | `worker/src/sync/graphql.ts` | GitHub GraphQL client (`fetch()`) |
 | `worker/src/sync/queries.ts` | GraphQL query strings |
 | `worker/src/sync/parse.ts` | Issue/edge parsing |
